Improve performance of log splitter
Removes warning when encountering unsplittable value that is too large,
since it requires calculating the size of every non-array or non-object,
which is incredibly expensive.

Iterates over collections using numeric for loops instead of map/reduce.

Calculate object size using `Buffer.byteLength` instead of splitting the
stringified value when possible. This resulted in approx 2x speed.

Overall, the performance difference is approx. 10x when called on a
large object.